requirements.txt
fastapi~=0.111.0
uvicorn~=0.29.0
celery~=5.4.0
redis~=5.0.4
openai
pydantic~=2.7.0
python-dotenv~=1.0.1
ffmpeg-python~=0.2.0
OpenTimelineIO-Plugins~=0.17.0
Pillow~=11.3.0
yt-dlp
SQLAlchemy~=2.0.30
google-auth~=2.29.0
requests
rich